# docs/swagger QuickOrder.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "QuickOrder.Api",
        "version": "v1"
    },
    "paths": {
...
        "/api/webhook/receiver/mercadopago": {
            "post": {
                "tags": [
                    "MercadoPagoWebHookReceiver"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebHookData"
                            }
                        },
                        "text/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebHookData"
                            }
                        },
                        "application/*+json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebHookData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "text/plain": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            },
                            "text/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Data": {
                "type": "object",
                "properties": {
                    "Id": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": false
            },
...
            "ProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "nullable": true
                    },
                    "title": {
                        "type": "string",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32",
                        "nullable": true
                    },
                    "detail": {
                        "type": "string",
                        "nullable": true
                    },
                    "instance": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "additionalProperties": {}
            },
...
            "WebHookData": {
                "type": "object",
                "properties": {
                    "Action": {
                        "type": "string",
                        "nullable": true
                    },
                    "Api_version": {
                        "type": "string",
                        "nullable": true
                    },
                    "Data": {
                        "$ref": "#/components/schemas/Data"
                    },
                    "Date_created": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "Id": {
                        "type": "string",
                        "nullable": true
                    },
                    "Live_mode": {
                        "type": "boolean"
                    },
                    "Type": {
                        "type": "string",
                        "nullable": true
                    },
                    "User_id": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "additionalProperties": false
            }
        }
    }
}